Initialize and refactor coordinator code
Initialize coordinator with two background workers: coordinator & auxiliary
