+++ src/coordinator/plc_coordinator.h
@@ -12,27 +12,27 @@
  * protocol: CO_PROTO_TCP, CO_PROTO_UNIX, CO_PROTO_UNIXPACKET
  * address: TCP(host:port), UNIX(/path/to/file.socket)
  */
-typedef enum {
+typedef enum CoordinatorState {
     CO_STATE_UNINITIALIZED = 1,
     CO_STATE_READY,
     CO_STATE_EXITING,
 } CoordinatorState;
 
-typedef enum {
+typedef enum CoordinatorProtocol {
     CO_PROTO_TCP = 1,
     CO_PROTO_UNIX,
     CO_PROTO_PROTO_UNIXPACKET,
-} coordinator_protocol;
+} CoordinatorProtocol;
 
-typedef struct
+typedef struct CoordinatorStruct
 {
     volatile CoordinatorState state;
-    coordinator_protocol protocol;
+    CoordinatorProtocol protocol;
     char address[504];
 } CoordinatorStruct;
 #define CO_SHM_KEY  "plcoordinator_shm"
 
-typedef struct
+typedef struct requester_info_entry
 {
     int id; // pid of QE
     int sock; // socket file descriptor between QE and coordinator
